+++ settings.js
@@ -5,7 +5,6 @@ const QUOTE_IMG   = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/
 
 // Compliance 32x
 const C32_IMG           = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Discord-Bot/master/images/Logos/Compliance%2032x.png';
-const C32_COLOR         = '#76C945';
 const C32_ID            = '773983706582482946';
 const C32_MODERATOR_ID  = '773984348860711003';
 const C32_LOGS          = '798893497204998161'; // #logs channel
@@ -33,26 +32,22 @@ const C64_AUTOPUSH_FAIL = '797780954626261022'; // #council-talk
 
 // Compliance Addons
 const CADDONS_IMG          = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Discord-Bot/master/images/Logos/Compliance%20Addons.png';
-const CADDONS_COLOR        = '#6D6163';
 const CADDONS_ID           = '614160586032414845';
 const CADDONS_MODERATOR_ID = '614163845656608774';
 
 // Compliance Tweaks
 const CTWEAKS_IMG          = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Discord-Bot/master/images/Logos/Compliance%20Tweaks.png';
-const CTWEAKS_COLOR        = '#658430';
 const CTWEAKS_ID           = '720966967325884426';
 const CTWEAKS_COUNTER      = '750638888296382504'; // Member counter channel
 const CTWEAKS_MODERATOR_ID = '788108550848512071';
 
 // Compliance Mods
 const CMODS_IMG          = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Discord-Bot/master/images/Logos/Compliance%20Mods.png';
-const CMODS_COLOR        = '#349687';
 const CMODS_ID           = '748264625962877019';
 const CMODS_MODERATOR_ID = '748270445534380163';
 
 // Compliance Dungeons
 const CDUNGEONS_IMG          = 'https://raw.githubusercontent.com/Compliance-Resource-Pack/Discord-Bot/master/images/Logos/Compliance%20Dungeons.png';
-const CDUNGEONS_COLOR        = '#DF7C2B';
 const CDUNGEONS_ID           = '714910830272970834';
 const CDUNGEONS_MODERATOR_ID = '714911578750844988';
 const CDUNGEONS_SUBMIT       = '715236892945285181';
@@ -63,7 +58,6 @@ module.exports = {
 	QUOTE_IMG,
 
 	C32_IMG,
-	C32_COLOR,
 	C32_ID,
 	C32_LOGS,
 	C32_MOD_LOGS,
@@ -89,23 +83,19 @@ module.exports = {
 	C64_AUTOPUSH_FAIL,
 
 	CADDONS_IMG,
-	CADDONS_COLOR,
 	CADDONS_ID,
 	CADDONS_MODERATOR_ID,
 
 	CTWEAKS_IMG,
-	CTWEAKS_COLOR,
 	CTWEAKS_ID,
 	CTWEAKS_COUNTER,
 	CTWEAKS_MODERATOR_ID,
 
 	CMODS_IMG,
-	CMODS_COLOR,
 	CMODS_ID,
 	CMODS_MODERATOR_ID,
 
 	CDUNGEONS_IMG,
-	CDUNGEONS_COLOR,
 	CDUNGEONS_ID,
 	CDUNGEONS_MODERATOR_ID,
 	CDUNGEONS_SUBMIT